refactor: minor changes, can't click accept changes without selecting an option
