+++ src/app/components/dialogs/mapSettingsDialog/mapSettingsDialog.component.scss
@@ -11,7 +11,7 @@
   align-items: center;
   justify-content: space-evenly;
   .color-box {
-    width: 90%;
+    width: 89%;
     padding: 20px;
   }
   .colour-option {
