+++ src/app/pages/quickMaps/components/colourObjects/colourGradient.ts
