+++ src/app/pages/quickMaps/components/colourObjects/colourPalette.ts
