+++ src/app/pages/quickMaps/components/colourObjects/colourPaletteType.enum.ts
